requirements.txt
absl-py>=1.0.0  # Upgrade to a more recent version if available
astor>=0.8.1    # Verify if a newer version is available
cachetools>=5.0.0  # Update to the latest version
certifi>=2020.12.5  # Update to the latest version
chardet>=4.0.0   # Note: `chardet` has been replaced by `charset-normalizer` in some newer projects
click>=8.0.0   # Update to the latest version
Flask>=2.0.0   # Flask 2.0 and above supports Python 3.12
gast>=0.4.0    # Update if necessary
google-auth>=2.0.0  # Update to the latest version
google-auth-oauthlib>=0.4.4  # Update to the latest version
google-pasta>=0.2.0  # Update if necessary
grpcio>=1.39.0  # Update to the latest version
gunicorn>=20.1.0  # Update to the latest version
h5py>=3.6.0  # Ensure compatibility with TensorFlow version
idna>=3.3    # Update to the latest version
importlib-metadata>=4.0  # This might not be needed if Python 3.12 natively supports metadata handling
itsdangerous>=2.0.0  # Compatible with newer Flask versions
Jinja2>=3.0.0   # Update to the latest version
joblib>=1.1.0   # Update to the latest version
Keras-Applications>=1.0.8  # Ensure compatibility with TensorFlow version
Keras-Preprocessing>=1.1.2  # Ensure compatibility with TensorFlow version
Markdown>=3.3.4  # Update to the latest version
MarkupSafe>=2.0.0  # Compatible with newer Flask versions
numpy>=1.21.0  # Update to the latest version, check for compatibility with TensorFlow and other packages
oauthlib>=3.2.0  # Update to the latest version
opt-einsum>=3.3.0  # Ensure compatibility with TensorFlow version
pickle-mixin>=1.0.2  # Ensure compatibility
Pillow>=9.0.0  # Update to the latest version
protobuf>=3.19.0  # Ensure compatibility with TensorFlow and other packages
pyasn1>=0.4.8  # Update if necessary
pyasn1-modules>=0.2.8  # Update if necessary
requests>=2.26.0  # Update to the latest version
requests-oauthlib>=1.3.1  # Update to the latest version
rsa>=4.8  # Update if necessary
scikit-learn>=1.0.0  # Update to the latest version, ensure compatibility with other packages
scipy>=1.7.0  # Update to the latest version
six>=1.16.0  # Ensure compatibility
sklearn==0.0  # This package is a placeholder; use `scikit-learn`
tensorboard>=2.8.0  # Ensure compatibility with TensorFlow version
tensorflow>=2.8.0  # TensorFlow 2.8 and later supports Python 3.12
tensorflow-estimator>=2.8.0  # Ensure compatibility with TensorFlow version
termcolor>=1.1.0  # Update if necessary
threadpoolctl>=3.0.0  # Ensure compatibility with scikit-learn
urllib3>=1.26.7  # Update to the latest version
Werkzeug>=2.0.0  # Compatible with Flask 2.0 and newer
wrapt>=1.13.0  # Ensure compatibility
zipp>=3.6.0  # Update to the latest version
